requirements.txt
Django
django-cors-headers
djangorestframework
flake8-django
mysqlclient
python-decouple
autopep8